+++ CHANGELOG.md
@@ -5,6 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.5 - 2020-01-13
+### Fixed
+-   Use $_SERVER instead of filter_input(INPUT_SERVER, ...) as not working in PHP versions with enabled FastCGI
+
+## 0.2.4 - 2020-01-13
+### Added
+-   Webhook::getHmacSha256() method visibility changed from private to public
+
+## 0.2.3 - 2019-12-04
+### Added
+-   PHPDoc for ClientInterface, it was unclear before that client method may throw exceptions
+
+## 0.2.2 - 2019-12-03
+### Added
+-   Allow `.myshopify.com` to be included in shop's handle passed to ClientConfig
+
+### Fixed
+-   Throw Guzzle request exception right away if it doesn't have response object to handle
+
+### Removed
+-   `vlucas/phpdotenv` dependency, moved to `require-dev`
+
+## 0.2.1 - 2019-11-24
+### Fixed
+-   Call to a member function on null fix in RequestException
+-   Autoload file with functions
+
 ## 0.2.0 - 2019-11-22
 ### Added
 -   Client class added as a replacement of the API class 
@@ -14,6 +41,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
 -   API class deprecated, use Client instead
+
+## 0.1.4 - 2019-11-24
+### Fixed
+-   Call to a member function on null fix in RequestException
 
 ## 0.1.3 - 2019-11-19
 ### Added